feat: multi-chunk message handling

+++ LiveStreamingServer/Rtmp/Core/RtmpMessagetHandler/Headers/ChunkMessageHeader.cs
@@ -1,17 +1,16 @@
 ﻿using LiveStreamingServer.Newtorking;
 using LiveStreamingServer.Newtorking.Contracts;
 using LiveStreamingServer.Rtmp.Core.Extensions;
-using System.Reflection.PortableExecutable;
 
 namespace LiveStreamingServer.Rtmp.Core.RtmpMessageHandler.Headers
 {
-    public record struct ChunkMessageHeaderType0(uint Timestamp, uint MessageLength, byte MessageTypeId, uint MessageStreamId)
+    public record struct ChunkMessageHeaderType0(uint Timestamp, int MessageLength, byte MessageTypeId, uint MessageStreamId)
     {
         public static async Task<ChunkMessageHeaderType0> ReadAsync(INetBuffer netBuffer, ReadOnlyNetworkStream networkStream, CancellationToken cancellationToken)
         {
             await netBuffer.CopyStreamData(networkStream, 11, cancellationToken);
             var timestampDelta = netBuffer.ReadUInt24BigEndian();
-            var messageLength = netBuffer.ReadUInt24BigEndian();
+            var messageLength = (int)netBuffer.ReadUInt24BigEndian();
             var messageTypeId = netBuffer.ReadByte();
             var messageStreamId = netBuffer.ReadUInt32();
 
@@ -19,14 +18,14 @@ namespace LiveStreamingServer.Rtmp.Core.RtmpMessageHandler.Headers
         }
     }
 
-    public record struct ChunkMessageHeaderType1(uint TimestampDelta, uint MessageLength, byte MessageTypeId)
+    public record struct ChunkMessageHeaderType1(uint TimestampDelta, int MessageLength, byte MessageTypeId)
     {
         public static async Task<ChunkMessageHeaderType1> ReadAsync(INetBuffer netBuffer, ReadOnlyNetworkStream networkStream, CancellationToken cancellationToken)
         {
             await netBuffer.CopyStreamData(networkStream, 7, cancellationToken);
 
             var timestampDelta = netBuffer.ReadUInt24BigEndian();
-            var messageLength = netBuffer.ReadUInt24BigEndian();
+            var messageLength = (int)netBuffer.ReadUInt24BigEndian();
             var messageTypeId = netBuffer.ReadByte();
 
             return new ChunkMessageHeaderType1(timestampDelta, messageLength, messageTypeId);
